requirements.txt
Flask==3.1.0
matplotlib==3.10.1
mlflow==2.22.0
mlflow_skinny==2.22.0
numpy==2.2.5
pandas==2.2.3
Pillow==11.2.1
Requests==2.32.3
scikit_learn==1.6.1
torch==2.7.0+cpu
